refactor: move GlobalExceptionHandler.kt to advice dir
