--- src/main/kotlin/ecommerce/GlobalExceptionHandler.kt
@@ -1,16 +1,0 @@
-package ecommerce
-
-import org.springframework.dao.EmptyResultDataAccessException
-import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.ExceptionHandler
-import org.springframework.web.bind.annotation.RestControllerAdvice
-
-@RestControllerAdvice
-class GlobalExceptionHandler {
-    @ExceptionHandler(EmptyResultDataAccessException::class)
-    fun handleEmptyResult(ex: EmptyResultDataAccessException): ResponseEntity<String> {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-            .body("Resource not found: ${ex.message}")
-    }
-}
